[sailfish-crypto] Remove deleteStoredSecret from Crypto plugin interface.  Contributes to JB#40058
It is now a requirement that any CryptoPlugin which supports storing
keys MUST implement the EncryptedStoragePlugin interface, and keys
are stored as "normal" secrets in its encrypted storage database.

This allows us to unify the implementation of deleteStoredKey() by
ensuring that the key will be deleted via the deleteSecret() codepath.