[sailfish-crypto] Unify stored-crypto-key book-keeping. Contributes to JB#40058
Any CryptoPlugin which supports storing crypto keys (i.e. via
generateStoredKey()) now MUST also implement the EncryptedStoragePlugin
interface.  This allows us to unify the way in which key-related
information is stored in the master (bookkeeping) table.

Any Key which is stored via a generateStoredKey() request will now have
its metadata stored as for any other Secret.  This allows us to enforce
strong foreign key constraints on these entries, preventing consistency
issues.

In order to achieve this, we add some helper methods in the Secrets
implementation allowing us to store just bookkeeping information for
a key (secret) which will be generated and stored internally by the
plugin (meaning that the secret key data will never enter the daemon
process' address space), and another such method to delete it (for
the cleanup case, as the "normal" full-secret-deletion codepath will
be called in the deleteStoredKey()/deleteSecret() case).