[sailfish-secrets] Expose Secret type in top-level API. Contributes to JB#36797
This commit improves the Secret type and exposes it in the top-level
API.  An instance of Secret may contain a name, collectionName and
secret data, as well as a type tag.  Each Secret is identified by an
identifier which consists of the name of the Secret and the name of
the collection in which the Secret is stored.

The Secret and Secret::Identifier types are then used throughout the
daemon implementation instead of separate name/collectionName/data
fields.