Rename secretName to hashedSecretName where appropriate in plugin API
The extension plugins API often takes a secret name as a parameter
where the value is a multiply-hashed version of the original secret
name (to provide some protection against known-plain-text attacks).

This commit renames the secretName parameter to hashedSecretName
where appropriate, and also adds an encryptedSecretName parameter
so that this can be retrieved and decrypted so as to be able to
return the original plaintext name to clients where necessary.