[sailfish-secrets] Allow EncryptedStorage plugins to be Crypto plugins. Contributes to JB#36797
Secrets plugins and Crypto plugins are installed to separate
directories, and thus previously a plugin could not be both a
Secrets plugin and also a Crypto plugin.

This commit ensures that we check whether a particular EncryptedStorage
plugin also implements the Crypto plugin interface, allowing it to
implement the generateStoredKey() method securely (i.e. without the
secret key data ever entering the daemon process address space).