Fix DBus serialisation of filter data, use field serialisation of Key
This commit ensures that we serialise filter data as a variant map,
to ensure that the introspection signature matches the serialisation
format.  It also expands the DBus serialisation of the Key type so
that it is serialised field-wise (with a complex and expanded DBus
type serialisation signature) to avoid using the internal QByteArray
serialisation and deserialisation methods for DBus marshalling, as
this will allow us to implement serialisation and deserialisation in
raw C code without requiring linking against Qt for QDataStream etc.

It also fixes a couple of errors in the introspection XML data.