[sailfish-secrets] Add framework overview documentation. Contributes to JB#36797
This commit adds documentation for the Sailfish OS Secrets and Crypto
Framework, including overview documentation for the Sailfish OS Secrets
Library and the Sailfish OS Crypto Library.

It also changes the install path of C++ plugins and devel headers
for symmetry with the install path for QML plugins and documentation.

Contributes to JB#36797