[sailfish-crypto] Allow the client to limit retrieve key data. Contributes to JB#40058
Previously, the StoredKeyRequest would retrieve all of the data
associated with a key entry, including metadata, public key data, and
private key data.  This commit allows the client to specify which
of these portions of data they wish to retrieve.

This allows, for example, the client to retrieve just the public
key data portion, allowing the client to limit potential secret
key data leakage in the case where they suspect that the secrecy
of their process's memory address space may be compromised.

Contributes to JB#40058