Don't update the status of a request without a valid result
In some circumstances, a QDBusPendingReply which was just created
via asyncCallWithArgumentList() or asyncCall() can report that it
isFinished(), despite not having yet been delivered.

This commit ensures we work around this issue, by checking to see
if the reply has a valid error code set in it (an immediate,
synchronous failure always results in a valid error code being
set) and if not, connecting an asynchronous watcher to it as per
normal, instead of immediately transitioning the request status
to Finished.