Don't attempt to read secret key data into daemon unnecessarily
In the case where the client is attempting to perform an operation
such as encrypt/decrypt/sign/verify, and where the provided key
is a key reference (i.e., key containing just an identifier),
we previously read that key from where it was stored, and then
passed that key to the crypto plugin for use in the operation.

However, if the crypto plugin is also the storage plugin for that
key, then we should immediately pass the given (reference) key
to the crypto plugin as the key to use in the operation, thus
avoiding reading the secret key data into the daemon unnecessarily.