Refactor InteractionRequest into InteractionParameters
This will allow us to use the InteractionParameters as a user-supplied
argument in the future (e.g. when accepting user-input directly
to be stored as a secret, or used as input key data from which
a key may be derived).

This also changes the interface of the AuthenticationPlugin
to clearly separate the two use-cases:
1) authenticate user (verify identity)
2) receive user input (e.g. passphrase etc)