[sailfish-crypto] Include explicit signature parameter to verify(). Contributes to JB#40058
Previously the verify() function took a single QByteArray "data"
parameter, which was intended to contain both the signature and the
signed data, which would require the plugin to unpact it appropriately
for the given padding and digest function.

This commit separates that into two parameters: signature and data,
providing a nicer API for both clients and plugins.