[sailfish-secrets] Allow NoAccessControl semantic. Contributes to JB#36797
This commit adds support for the "NoAccessControl" semantic, whereby
a collection (or standalone secret) with that semantic applied may
be accessed by any application (not just the owner application),
without any system-mediated access control applied.

This type of access control is not recommended unless the locking
semantic is CustomLockAccessRelock, in order to force the user to
explicitly permit every access from every application.

Contributes to JB#36797