[sailfish-secrets] Implement master lock functionality. Contributes to JB#36797
This commit adds support for encrypting device-locked collections
with a key derived from a master-passphrase.  Another key derived from
the master-passphrase is used to encrypt the bookkeeping database.

It also adds support for LockCodeRequest type which allows the
user to provide the passphrase (e.g. after the device has booted,
in order to unlock the secrets and crypto service), or to change
the passphrase (which will cause any device-locked data to be
re-encrypted with the key derived from the new passphrase).

Finally, it allows plugins to be unlocked with the generated key,
for those cases where a plugin requires a key to be provided
before its functionality is made available.

Contributes to JB#36797