[sailfish-secrets] Derive key from code for secrets encryption. Contributes to JB#36797
Previously, we passed the authentication code from the user to the
secrets encryption (or encryptedstorage) plugin, and left it entirely
up to the plugin to turn that into a key.  This had the advantage
that the plugin would produce the correct key size for its encryption
or decryption operation, but had the disadvantage that we then did
not handle the device lock code in the same way.

This commit unifies the handling, by adding an explicit key derivation
step to convert user-provided authentication codes into keys, where
required.

Also improve code-reuse of the key derivation function from the
OpenSSL EVP routines wrapper.

Contributes to JB#36797