[sailfish-secrets] Allow plugin-target lock code request. Contributes to JB#36797
Previously, we only supported the bookkeeping database, a collection,
or a standalone secret to be the target of a code code request.
This commit adds support for specifying an extension plugin as the
target of the request, in case a particular plugin needs to be
explicitly unlocked via a non-masterlock (bookkeeping database key)
lock code request.

Contributes to JB#36797