[sailfish-secrets] Serialise plugin operations via QtConcurrent. Contributes to JB#36797
We define two thread pools, each of which contains a single thread,
which are used to run functions exposed by plugins of the Secrets
and Crypto domains respectively.

This ensures:
1) the main thread will not be blocked synchronously executing
   long cryptographic operations provided by client plugins
2) client plugins don't have to manage operation serialisation
   as they are guaranteed to be serialised and non-interfering

This commit further then ensures that the secrets I/O thread pool
is used to perform plugin operations for secrets API requests.

Contributes to JB#36797