[sailfish-secrets] Prevent bookkeeping database operation interleaving. Contributes to JB#36797
Now that plugin operations are performed on separate threads, it is
possible that a long-running plugin operation can block the plugin
thread while the main thread is still active to partially process
incoming client requests.

This creates the possibility that a given request within a larger
sequence of requests (such as create collection followed by delete
collection) might be partially completed on the main thread before the
prior request has finished completely.

This commit ensures that we prevent such request interleaving on the
main thread when operating on the same collection.

Contributes to JB#36797