[sailfish-secrets] Write command line tool. Contributes to JB#36797
This commit adds a command line tool which allows the user to perform
a variety of operations via the Secrets and Crypto APIs.

It can be used to manually test the various authentication and user
input flows, to manually verify that a particular plugin is working
as expected, or to perform useful operations (such as signing,
verification, encryption or decryption) via a particular plugin.

Contributes to JB#36797 and JB#40058