Ensure that collections are relocked as required by the unlock semantic
Previously, we would leave the collection unlocked after performing
an operation upon it.  This commit ensures that we appropriately
relock the collection if required by the unlock semantic.

Now that we enforce locking semantics more strictly, a variety of
unit tests needed to be fixed, to ensure they enforce the appropriate
behaviour.