[secrets] Split passphrase prompt text into multiple strings. Contributes to JB#40880
Typically this will be a message describing the reason for the prompt
and an 'instruction' describing the what to enter.

For entering new passphrases there are separate 'newInstruction'
and 'repeatInstruction' prompts for the entry with confirmation,
the 'instruction' prompt in reserved for entering an existing
passphrase in the event the prompt is for changing a passphrase.