[sailfish-secrets] Use auth_self for VerifyUser polkit policy. Contributes to JB#36797
This change ensures that we use the auth_self mode for polkit policy
authentication so that if the device user does NOT have a lock code
set, the authentication still succeeds.

It also means that even if the user does have a lock code set, the
user will be able to Verify with just a confirm (yes/no) prompt,
rather than having to re-enter their lock code.

If we wish to change that behaviour in the future, we can add a new
policy case to require lock code entry.