[sailfish-secrets] Use appropriate auth plugin depending on permissions. Contributes to JB#36797
If a collection is created by a privileged application with its
authentication plugin set to the in-app authentication plugin
(e.g. for in-process flows in Settings application), we don't want
non-privileged third-party applications to be able to use in-app
authentication to unlock the collection (as that would leak the
unlock code to the third-party application).

This commit ensures that we always use the appropriate (system default)
authentication plugin for non-privileged applications, if they do
not own the collection (or standalone secret) in question.

Contributes to JB#36797