[sailfish-crypto] Trigger collection unlock flow for crypto operations. Contributes to JB#40058
If a collection which contains a key is currently locked (due to the
unlock semantic being AccessRelock, for example), then if a client
attempts to use that key (e.g. to encrypt/decrypt/sign/verify) the
user should be prompted to unlock the collection in order to allow
the key to be retrieved in order to perform the operation.

This commit ensures that this occurs, where previously the operations
would have failed with a CollectionLocked error.

Contributes to JB#40058