[sailfish-secrets] Add support for keep-unlocked semantics to secrets-tool. Contributes to JB#36797
This commit adds support for specifying the lock semantic which should
be applied to the collection or standalone secret being created
with the secrets-tool.

It also adds a new manual test script which performs operations on
a "keep-unlocked" collection, to test that the semantic is applied
appropriately (and so the user is not prompted to enter the unlock
code for every operation, as they are for manual-test.sh).

Finally, it supports specifying secret data via the tool rather
than requiring that data to be entered via UI prompt.

Contributes to JB#36797