[sailfish-secrets] Initialize passwordagent once accepted. Contributes to JB#36797
We build both a "normal" and a "test" version of various plugins.
Upon daemon startup, we load all plugins, and then unload any which
don't match the current mode (autotest = true/false).

This has the unfortunate side effect that the test version of the
passwordagent plugin will overwrite the socket file originally
created by the normal version of the password agent plugin, before
being unloaded.

This commit ensures that we allow plugins to perform initialization
in a specific initialize() method which is only called once the
plugin has been "accepted" for use by the daemon according to the
current mode.

Contributes to JB#36797