[sailfish-secrets] Fallback to hardcoded p2p address on discovery failure. Contributes to JB#42034
If a client service is running as root and cannot access the user
session bus to perform discovery, a connection to the secretsd service
previously would not have been established.

This commit adds a fallback address which should match the real address
to allow such services to make use of the secretsd service.

Contributes to JB#42034