[sailfish-secrets] Fix master lock key initialization. Contributes to JB#42034
If the default crypto plugin was not installed to the device, the daemon
would previously fail to correctly generate the master lock key and
device lock key, which would cause a variety of operations to fail.

This commit ensures that key initialization will search for:
 - plugin specified by SAILFISH_SECRETSD_MASTERLOCK_CRYPTOPLUGIN
 - default crypto plugin
 - any plugin which can perform key generation
in that order, so that appropriate keys can be generated.

It also ensures that the key generation (and test cipher generation)
operations are performed on the appropriate thread.

Finally, it stores the name of the plugin used to generate the key
to the test cipher data file.

Contributes to JB#42034